Added logic to redirect user to landing page if on specific URLs
